Update AutoScreenshot.py
There was a necessary change to accommodate a new file naming system on mac for screenshots. Basically I didn't know this but there is an added bit of code concerning something to do with a header for the file. But either way, I dealt with it now as well as fixing there not being enough time to press enter after loading the file name in the upload file window